refactor(core): modularizar soft delete en paquete dedicado con compatibilidad
