+++ core/soft_delete/__init__.py
@@ -1,0 +1,9 @@
+"""Core soft-delete domain package."""
+
+from .base import SoftDeleteManager, SoftDeleteModelMixin, SoftDeleteQuerySet
+
+__all__ = [
+    "SoftDeleteQuerySet",
+    "SoftDeleteManager",
+    "SoftDeleteModelMixin",
+]
