+++ core/soft_delete/signals.py
@@ -1,0 +1,11 @@
+"""Domain signals emitted for soft-delete lifecycle events."""
+
+from django.dispatch import Signal
+
+# Sent after a model instance was soft-deleted.
+# kwargs: user, cascade, root
+post_soft_delete = Signal()
+
+# Sent after a model instance was restored.
+# kwargs: user, cascade, root
+post_restore = Signal()
